web/requirements.txt
Flask
Flask-WTF
Jinja2
Flask-SQLAlchemy
boto3
psycopg2
